+++ app/src/main/java/ru/practicum/android/diploma/di/StorageModule.kt
@@ -1,7 +1,6 @@
 package ru.practicum.android.diploma.di
 
 import android.content.Context
-import android.content.SharedPreferences
 import com.google.gson.Gson
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
